(#965) Remove need for current.jst

+++ gobblin-metastore/src/main/java/gobblin/metastore/FsStateStore.java
@@ -24,20 +24,25 @@ import java.net.URI;
 import java.util.Collection;
 import java.util.List;
 
+import lombok.extern.slf4j.Slf4j;
 import org.apache.hadoop.conf.Configuration;
 import org.apache.hadoop.fs.FileStatus;
 import org.apache.hadoop.fs.FileSystem;
 import org.apache.hadoop.fs.Path;
+import org.apache.hadoop.fs.PathFilter;
 import org.apache.hadoop.fs.permission.FsPermission;
 import org.apache.hadoop.io.SequenceFile;
 import org.apache.hadoop.io.Text;
 import org.apache.hadoop.io.compress.DefaultCodec;
 
+import com.google.common.base.Preconditions;
 import com.google.common.base.Strings;
+import com.google.common.collect.ImmutableList;
 import com.google.common.collect.Lists;
 import com.google.common.io.Closer;
 
 import gobblin.configuration.State;
+import gobblin.metastore.util.StateStoreTableInfo;
 import gobblin.util.HadoopUtils;
 
 
@@ -58,6 +63,7 @@ import gobblin.util.HadoopUtils;
  *
  * @author Yinan Li
  */
+@Slf4j
 public class FsStateStore<T extends State> implements StateStore<T> {
 
   public static final String TMP_FILE_PREFIX = "_tmp_";
@@ -72,45 +78,77 @@ public class FsStateStore<T extends State> implements StateStore<T> {
   // Class of the state objects to be put into the store
   private final Class<T> stateClass;
 
-  public FsStateStore(String fsUri, String storeRootDir, Class<T> stateClass) throws IOException {
-    this.conf = new Configuration();
-    this.fs = FileSystem.get(URI.create(fsUri), this.conf);
-    this.useTmpFileForPut = !FS_SCHEMES_NON_ATOMIC.contains(this.fs.getUri().getScheme());
-    this.storeRootDir = storeRootDir;
-    this.stateClass = stateClass;
+  // The extension of the state file
+  private final String extension;
+
+  public FsStateStore(String fsUri, String storeRootDir, Class<T> stateClass, String extension) throws IOException {
+    this(FileSystem.get(URI.create(fsUri), new Configuration()), storeRootDir, stateClass, extension);
   }
 
-  public FsStateStore(FileSystem fs, String storeRootDir, Class<T> stateClass) {
+  public FsStateStore(String storeUrl, Class<T> stateClass, String extension) throws IOException {
+    this(new Path(storeUrl), stateClass, extension);
+  }
+
+  public FsStateStore(Path storePath, Class<T> stateClass, String extension) throws IOException {
+    this(storePath.getFileSystem(new Configuration()), storePath.toUri().getPath(), stateClass, extension);
+  }
+
+  public FsStateStore(FileSystem fs, String storeRootDir, Class<T> stateClass, String extension) throws IOException {
     this.fs = fs;
     this.useTmpFileForPut = !FS_SCHEMES_NON_ATOMIC.contains(this.fs.getUri().getScheme());
     this.conf = this.fs.getConf();
     this.storeRootDir = storeRootDir;
     this.stateClass = stateClass;
+    this.extension = normalizeExtension(extension);
   }
 
-  public FsStateStore(String storeUrl, Class<T> stateClass) throws IOException {
-    this.conf = new Configuration();
-    Path storePath = new Path(storeUrl);
-    this.fs = storePath.getFileSystem(this.conf);
-    this.useTmpFileForPut = !FS_SCHEMES_NON_ATOMIC.contains(this.fs.getUri().getScheme());
-    this.storeRootDir = storePath.toUri().getPath();
-    this.stateClass = stateClass;
-  }
-
+  /**
+   * Create a new store.
+   *
+   * <p>
+   *     A store that does not exist will be created when any put
+   *     method is called against it.
+   * </p>
+   *
+   * @param storeName store name
+   * @return if the store is successfully created
+   * @throws IOException
+   */
   @Override
   public boolean create(String storeName) throws IOException {
+    Preconditions.checkArgument(!Strings.isNullOrEmpty(storeName), "Store name is null or empty.");
+
     Path storePath = new Path(this.storeRootDir, storeName);
     return this.fs.exists(storePath) || this.fs.mkdirs(storePath, new FsPermission((short) 0755));
   }
 
+  /**
+   * Create a new table in a store.
+   *
+   * <p>
+   *     A table that does not exist will be created when any put
+   *     method is called against it.
+   * </p>
+   *
+   * @param storeName store name
+   * @param tableName table name
+   * @return if the table is successfully created
+   * @throws IOException
+   */
   @Override
   public boolean create(String storeName, String tableName) throws IOException {
+    Preconditions.checkArgument(!Strings.isNullOrEmpty(storeName), "Store name is null or empty.");
+    Preconditions.checkArgument(!Strings.isNullOrEmpty(tableName), "Table name is null or empty.");
+    Preconditions.checkArgument(!isCurrentFile(tableName),
+        String.format("Table name is %s or %s%s", StateStoreTableInfo.CURRENT_NAME, StateStoreTableInfo.CURRENT_NAME,
+        this.extension));
+
     Path storePath = new Path(this.storeRootDir, storeName);
     if (!this.fs.exists(storePath) && !create(storeName)) {
       return false;
     }
 
-    Path tablePath = new Path(storePath, tableName);
+    Path tablePath = new Path(storePath, normalizeTableName(tableName));
     if (this.fs.exists(tablePath)) {
       throw new IOException(String.format("State file %s already exists for table %s", tablePath, tableName));
     }
@@ -118,13 +156,24 @@ public class FsStateStore<T extends State> implements StateStore<T> {
     return this.fs.createNewFile(tablePath);
   }
 
+  /**
+   * Check whether a given table exists.
+   *
+   * @param storeName store name
+   * @param tableName table name
+   * @return whether the given table exists
+   * @throws IOException
+   */
   @Override
   public boolean exists(String storeName, String tableName) throws IOException {
-    Path tablePath = new Path(new Path(this.storeRootDir, storeName), tableName);
-    return this.fs.exists(tablePath);
+    Preconditions.checkArgument(!Strings.isNullOrEmpty(storeName), "Store name is null or empty.");
+    Preconditions.checkArgument(!Strings.isNullOrEmpty(tableName), "Table name is null or empty.");
+
+    Path tablePath = getTablePath(storeName, tableName);
+    return tablePath != null && this.fs.exists(tablePath);
   }
 
-  /**
+    /**
    * See {@link StateStore#put(String, String, T)}.
    *
    * <p>
@@ -134,28 +183,14 @@ public class FsStateStore<T extends State> implements StateStore<T> {
    */
   @Override
   public void put(String storeName, String tableName, T state) throws IOException {
-    String tmpTableName = this.useTmpFileForPut ? TMP_FILE_PREFIX + tableName : tableName;
-    Path tmpTablePath = new Path(new Path(this.storeRootDir, storeName), tmpTableName);
+    Preconditions.checkArgument(!Strings.isNullOrEmpty(storeName), "Store name is null or empty.");
+    Preconditions.checkArgument(!Strings.isNullOrEmpty(tableName), "Table name is null or empty.");
+    Preconditions.checkArgument(!isCurrentFile(tableName),
+        String.format("Table name is %s or %s%s", StateStoreTableInfo.CURRENT_NAME, StateStoreTableInfo.CURRENT_NAME,
+        this.extension));
+    Preconditions.checkNotNull(state, "State is null.");
 
-    if (!this.fs.exists(tmpTablePath) && !create(storeName, tmpTableName)) {
-      throw new IOException("Failed to create a state file for table " + tmpTableName);
-    }
-
-    Closer closer = Closer.create();
-    try {
-      SequenceFile.Writer writer = closer.register(SequenceFile.createWriter(this.fs, this.conf, tmpTablePath,
-          Text.class, this.stateClass, SequenceFile.CompressionType.BLOCK, new DefaultCodec()));
-      writer.append(new Text(Strings.nullToEmpty(state.getId())), state);
-    } catch (Throwable t) {
-      throw closer.rethrow(t);
-    } finally {
-      closer.close();
-    }
-
-    if (this.useTmpFileForPut) {
-      Path tablePath = new Path(new Path(this.storeRootDir, storeName), tableName);
-      HadoopUtils.renamePath(this.fs, tmpTablePath, tablePath);
-    }
+    this.putAll(storeName, tableName, ImmutableList.of(state));
   }
 
   /**
@@ -168,7 +203,15 @@ public class FsStateStore<T extends State> implements StateStore<T> {
    */
   @Override
   public void putAll(String storeName, String tableName, Collection<T> states) throws IOException {
-    String tmpTableName = this.useTmpFileForPut ? TMP_FILE_PREFIX + tableName : tableName;
+    Preconditions.checkArgument(!Strings.isNullOrEmpty(storeName), "Store name is null or empty.");
+    Preconditions.checkArgument(!Strings.isNullOrEmpty(tableName), "Table name is null or empty.");
+    Preconditions.checkArgument(!isCurrentFile(tableName),
+        String.format("Table name is %s or %s%s", StateStoreTableInfo.CURRENT_NAME, StateStoreTableInfo.CURRENT_NAME,
+        this.extension));
+    Preconditions.checkNotNull(states, "States is null.");
+
+    String normalizedTableName = normalizeTableName(tableName);
+    String tmpTableName = this.useTmpFileForPut ? TMP_FILE_PREFIX + normalizedTableName : normalizedTableName;
     Path tmpTablePath = new Path(new Path(this.storeRootDir, storeName), tmpTableName);
 
     if (!this.fs.exists(tmpTablePath) && !create(storeName, tmpTableName)) {
@@ -189,15 +232,27 @@ public class FsStateStore<T extends State> implements StateStore<T> {
     }
 
     if (this.useTmpFileForPut) {
-      Path tablePath = new Path(new Path(this.storeRootDir, storeName), tableName);
+      Path tablePath = new Path(new Path(this.storeRootDir, storeName), normalizedTableName);
       HadoopUtils.renamePath(this.fs, tmpTablePath, tablePath);
     }
   }
 
+  /**
+   * Get all {@link State}s from a table.
+   *
+   * @param storeName store name
+   * @param tableName table name
+   * @return (possibly empty) list of {@link State}s from the given table
+   * @throws IOException
+   */
   @Override
   public T get(String storeName, String tableName, String stateId) throws IOException {
-    Path tablePath = new Path(new Path(this.storeRootDir, storeName), tableName);
-    if (!this.fs.exists(tablePath)) {
+    Preconditions.checkArgument(!Strings.isNullOrEmpty(storeName), "Store name is null or empty.");
+    Preconditions.checkArgument(!Strings.isNullOrEmpty(tableName), "Table name is null or empty.");
+    Preconditions.checkArgument(!Strings.isNullOrEmpty(stateId), "State id is null or empty.");
+
+    Path tablePath = getTablePath(storeName, tableName);
+    if (tablePath == null || !this.fs.exists(tablePath)) {
       return null;
     }
 
@@ -225,12 +280,29 @@ public class FsStateStore<T extends State> implements StateStore<T> {
     return null;
   }
 
+  /**
+   * Get a {@link State} with a given state ID from a the current table.
+   *
+   * @param storeName store name
+   * @param stateId state ID
+   * @return {@link State} with the given state ID or <em>null</em>
+   *         if the state with the given state ID does not exist
+   * @throws IOException
+   */
+  @Override
+  public T getCurrent(String storeName, String stateId) throws IOException {
+    return get(storeName, StateStoreTableInfo.CURRENT_NAME, stateId);
+  }
+
   @Override
   public List<T> getAll(String storeName, String tableName) throws IOException {
+    Preconditions.checkArgument(!Strings.isNullOrEmpty(storeName), "Store name is null or empty.");
+    Preconditions.checkArgument(!Strings.isNullOrEmpty(tableName), "Table name is null or empty.");
+
     List<T> states = Lists.newArrayList();
 
-    Path tablePath = new Path(new Path(this.storeRootDir, storeName), tableName);
-    if (!this.fs.exists(tablePath)) {
+    Path tablePath = getTablePath(storeName, tableName);
+    if (tablePath == null || !this.fs.exists(tablePath)) {
       return states;
     }
 
@@ -258,8 +330,22 @@ public class FsStateStore<T extends State> implements StateStore<T> {
     return states;
   }
 
+  /**
+   * Get all {@link State}s from the current table.
+   *
+   * @param storeName store name
+   * @return (possibly empty) list of {@link State}s from the current table
+   * @throws IOException
+   */
+  @Override
+  public List<T> getAllCurrent(String storeName) throws IOException {
+    return getAll(storeName, StateStoreTableInfo.CURRENT_NAME);
+  }
+
   @Override
   public List<T> getAll(String storeName) throws IOException {
+    Preconditions.checkArgument(!Strings.isNullOrEmpty(storeName), "Store name is null or empty.");
+
     List<T> states = Lists.newArrayList();
 
     Path storePath = new Path(this.storeRootDir, storeName);
@@ -275,32 +361,80 @@ public class FsStateStore<T extends State> implements StateStore<T> {
   }
 
   @Override
-  public void createAlias(String storeName, String original, String alias) throws IOException {
-    Path originalTablePath = new Path(new Path(this.storeRootDir, storeName), original);
-    if (!this.fs.exists(originalTablePath)) {
-      throw new IOException(String.format("State file %s does not exist for table %s", originalTablePath, original));
-    }
-
-    Path aliasTablePath = new Path(new Path(this.storeRootDir, storeName), alias);
-    Path tmpAliasTablePath = new Path(aliasTablePath.getParent(), new Path(TMP_FILE_PREFIX, aliasTablePath.getName()));
-    // Make a copy of the original table as a work-around because
-    // Hadoop version 1.2.1 has no support for symlink yet.
-    HadoopUtils.copyFile(this.fs, originalTablePath, this.fs, aliasTablePath, tmpAliasTablePath, true, this.conf);
-  }
-
-  @Override
   public void delete(String storeName, String tableName) throws IOException {
-    Path tablePath = new Path(new Path(this.storeRootDir, storeName), tableName);
-    if (this.fs.exists(tablePath)) {
+    Preconditions.checkArgument(!Strings.isNullOrEmpty(storeName), "Store name is null or empty.");
+    Preconditions.checkArgument(!Strings.isNullOrEmpty(tableName), "Table name is null or empty.");
+    Preconditions.checkArgument(!isCurrentFile(tableName),
+      String.format("Table name is %s or %s%s", StateStoreTableInfo.CURRENT_NAME, StateStoreTableInfo.CURRENT_NAME,
+      this.extension));
+
+    Path tablePath = getTablePath(storeName, tableName);
+    if (tablePath != null && this.fs.exists(tablePath)) {
       this.fs.delete(tablePath, false);
     }
   }
 
   @Override
   public void delete(String storeName) throws IOException {
+    Preconditions.checkArgument(!Strings.isNullOrEmpty(storeName), "Store name is null or empty.");
+
     Path storePath = new Path(this.storeRootDir, storeName);
     if (this.fs.exists(storePath)) {
       this.fs.delete(storePath, true);
     }
+  }
+
+  protected Path getTablePath(String storeName, String tableName) throws IOException {
+    StateStoreTableInfo tableInfo = StateStoreTableInfo.get(tableName, this.extension);
+    if (tableInfo.isCurrent()) {
+      return getLatestStateFilePath(storeName, tableInfo.getPrefix());
+    }
+    return new Path(new Path(this.storeRootDir, storeName), normalizeTableName(tableName));
+  }
+
+  private Path getLatestStateFilePath(String storeName, final String tableNamePrefix) throws IOException {
+    Path stateStorePath = new Path(this.storeRootDir, storeName);
+    if (!this.fs.exists(stateStorePath)) {
+      return null;
+    }
+
+    FileStatus[] stateStoreFileStatuses = this.fs.listStatus(stateStorePath, new PathFilter() {
+      @Override
+      public boolean accept(Path path) {
+        String name = path.getName();
+        return (Strings.isNullOrEmpty(tableNamePrefix) ||
+            name.startsWith(tableNamePrefix + StateStoreTableInfo.TABLE_PREFIX_SEPARATOR)) &&
+            name.endsWith(extension);
+      }
+    });
+
+    String latestStateFilePath = null;
+    for (FileStatus fileStatus : stateStoreFileStatuses) {
+      String tableName = fileStatus.getPath().getName();
+      if (latestStateFilePath == null) {
+        log.debug("Latest table for {}/{}* set to {}", storeName, tableNamePrefix, tableName);
+        latestStateFilePath = tableName;
+      } else if (tableName.compareTo(latestStateFilePath) > 0) {
+        log.debug("Latest table for {}/{}* set to {} instead of {}", storeName, tableNamePrefix, tableName, latestStateFilePath);
+        latestStateFilePath = tableName;
+      } else {
+        log.debug("Latest table for {}/{}* left as {}. Previous table {} is being ignored", storeName, tableNamePrefix, latestStateFilePath, tableName);
+      }
+    }
+
+    return latestStateFilePath == null ? null : new Path(new Path(this.storeRootDir, storeName), latestStateFilePath);
+  }
+
+  private boolean isCurrentFile(String tableName) {
+    StateStoreTableInfo tableInfo = StateStoreTableInfo.get(tableName, this.extension);
+    return tableInfo.isCurrent();
+  }
+
+  private String normalizeExtension(String extension) {
+    return extension.startsWith(".") ? extension : "." + extension;
+  }
+
+  private String normalizeTableName(String tableName) {
+    return tableName.endsWith(this.extension) ? tableName : tableName + this.extension;
   }
 }
